+++ CataloguesAnywhere/ModEntry.cs
@@ -4,6 +4,7 @@ using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewModdingAPI.Utilities;
 using StardewValley;
+using StardewValley.Menus;
 
 namespace CataloguesAnywhere;
 
@@ -47,17 +48,38 @@ internal class ModEntry : Mod
     /// <inheritdoc cref="IInputEvents.ButtonsChanged" />
     private void OnButtonsChanged(object? sender, ButtonsChangedEventArgs e)
     {
-        if (!Context.CanPlayerMove || !this.Config.Enabled)
+        if (!Context.IsWorldReady || !this.Config.Enabled)
             return;
 
         foreach ((string shopId, KeybindList keybind) in this.Config.Catalogues)
         {
             if (keybind.JustPressed())
             {
-                this.Monitor.Log($"Opening shop ID '{shopId}' per keybind '{keybind}'.");
-                Utility.TryOpenShopMenu(shopId, null as string);
+                if (this.CanOpenShopMenu(shopId))
+                {
+                    this.Monitor.Log($"Opening shop ID '{shopId}' per keybind '{keybind}'.");
+                    Utility.TryOpenShopMenu(shopId, null as string);
+                }
+
                 break;
             }
         }
+    }
+
+    /// <summary>Get whether a given shop menu can be opened now.</summary>
+    /// <param name="shopId">The shop ID being opened.</param>
+    private bool CanOpenShopMenu(string shopId)
+    {
+        return
+            // no menu open
+            Context.CanPlayerMove
+
+            // allow switching from one catalogue to another
+            || (
+                Game1.activeClickableMenu is ShopMenu { ShopId: not null } shopMenu
+                && shopMenu.ShopId != shopId
+                && this.Config.Catalogues.ContainsKey(shopMenu.ShopId)
+                && shopMenu.readyToClose()
+            );
     }
 }